Reorganize and add autoloader support
